Fix missing transient app items with CurrentWorkspaceOnly enabled
This fixes an issue that would only occur when CurrentWorkspaceOnly was
enabled. When a new app is opened and a transient item added, the item
is unconditionally marked as attached in update_visible_elements (in
DefaultApplicationDockItemProvider) when CurrentWorkspaceOnly is
disabled.
When it is enabled, however, the app item is not marked as attached
because it doesn't have any windows yet. When a window is subsequently
added during app startup, update_visible_elements was not called
previously, so the item would never be attached and hence stay
invisible (until update_visible_elements happened to called because of
some other event). This is now fixed by calling update_visible_elements
whenever a new window is added to an app item.

A related change (to avoid unnecessary calls to update_visible_elements)
is that ApplicationDockItem will not raise an app_window_added event
when the application state changes to running, which happens after a
window was already added.